优化多语言

+++ src/model/SystemAuth.php
@@ -56,10 +56,8 @@ class SystemAuth extends Model
      */
     public function onAdminDelete(string $ids)
     {
-        if (count($aids = str2arr($ids ?? '')) > 0) {
-            SystemNode::mk()->whereIn('auth', $aids)->delete();
-        }
-        sysoplog($this->oplogType, "删除{$this->oplogName}[{$ids}]及授权配置");
+        if (count($aids = str2arr($ids)) > 0) SystemNode::mk()->whereIn('auth', $aids)->delete();
+        sysoplog($this->oplogType, lang("删除%s[%s]及授权配置", [lang($this->oplogName), $ids]));
     }
 
     /**
